split out explorer into its own folder
