+++ uiCommon/explorer/relationTree.tsx
@@ -1,108 +1,13 @@
-import React, { useState } from "react";
-import useHashParam from "use-hash-param";
-import { Interpreter } from "../core/interpreter";
-import { Relation } from "../core/types";
-import * as styles from "./styles";
-import { RelationTable, TableCollapseState } from "./relationTable";
-import { noHighlight, HighlightProps } from "./term";
-import { useJSONLocalStorage, useBoolLocalStorage } from "./hooks";
-import { TreeView, TreeCollapseState, emptyCollapseState } from "./treeView";
-import { Tree, insertAtPath, filterTree } from "../util/tree";
-import { lastItem } from "../util/util";
+import React from "react";
+import { Relation } from "../../core/types";
+import { filterTree, insertAtPath, Tree } from "../../util/tree";
+import { lastItem } from "../../util/util";
+import { HighlightProps, noHighlight } from "../term";
+import { useBoolLocalStorage, useJSONLocalStorage } from "../hooks";
+import { emptyCollapseState, TreeCollapseState, TreeView } from "../treeView";
+import * as styles from "../styles";
 
-type RelationCollapseStates = { [key: string]: TableCollapseState };
-
-export function TabbedTables(props: { interp: Interpreter }) {
-  const allRules: Relation[] = Object.keys(props.interp.db.rules)
-    .sort()
-    .map((name) => ({ type: "Rule", name, rule: props.interp.db.rules[name] }));
-  const allTables: Relation[] = [
-    ...Object.keys(props.interp.db.tables)
-      .sort()
-      .map(
-        (name): Relation => ({
-          type: "Table",
-          name,
-          records: props.interp.db.tables[name],
-        })
-      ),
-    ...Object.keys(props.interp.db.virtualTables)
-      .sort()
-      .map(
-        (name): Relation => ({
-          type: "Table",
-          name,
-          records: props.interp.db.virtualTables[name](props.interp.db),
-        })
-      ),
-  ];
-  const allRelations: Relation[] = [...allTables, ...allRules];
-
-  const [highlight, setHighlight] = useState(noHighlight);
-  const highlightProps: HighlightProps = {
-    highlight,
-    setHighlight,
-    parentPaths: [],
-    childPaths: [],
-  };
-
-  const [curRelationName, setCurRelationName]: [
-    string,
-    (v: string) => void
-  ] = useHashParam(
-    "relation",
-    allRelations.length === 0 ? null : allRelations[0].name
-  );
-  const [
-    relationCollapseStates,
-    setRelationCollapseStates,
-  ] = useJSONLocalStorage<RelationCollapseStates>("collapse-state", {});
-
-  const curRelation = allRelations.find((r) => r.name === curRelationName);
-
-  return (
-    <div style={{ display: "flex" }}>
-      <div
-        style={{
-          border: "1px solid black",
-          overflow: "scroll",
-          paddingTop: 10,
-          paddingBottom: 10,
-          width: 275, // TODO: make the divider draggable
-        }}
-      >
-        <RelationTree
-          allRules={allRules}
-          allTables={allTables}
-          highlight={highlightProps}
-          curRelationName={curRelationName}
-          setCurRelationName={setCurRelationName}
-        />
-      </div>
-      <div style={{ padding: 10, border: "1px solid black", flexGrow: 1 }}>
-        {curRelation ? (
-          <RelationTable
-            relation={curRelation}
-            interp={props.interp}
-            highlight={highlightProps}
-            collapseState={relationCollapseStates[curRelationName] || {}}
-            setCollapseState={(st) =>
-              setRelationCollapseStates({
-                ...relationCollapseStates,
-                [curRelationName]: st,
-              })
-            }
-          />
-        ) : (
-          <em>No relations</em>
-        )}
-      </div>
-      <div>Visualizations here</div>
-    </div>
-  );
-}
-
-function RelationTree(props: {
+export function RelationTree(props: {
   allRules: Relation[];
   allTables: Relation[];
   curRelationName: string;
