+++ uiCommon/ide/errors.ts
@@ -1,7 +1,7 @@
 import { Interpreter } from "../../core/interpreter";
 import { Rec, Int } from "../../core/types";
 import { Span, dlToSpan } from "./types";
-import { treeFromRecords } from "../vizMappings/tree";
+import { treeFromRecords } from "../visualizations/tree";
 import { mapTree, filterTree, getLeaves } from "../../util/tree";
 
 // just `TypeError` is a builtin
