+++ uiCommon/relationTable.tsx
@@ -7,7 +7,7 @@ import { RuleC } from "./rule";
 import { makeTermWithBindings } from "../core/traceTree";
 import { TermView, noHighlight, HighlightProps } from "./term";
 import { TraceView } from "./trace";
-import { canTreeViz, treeFromRecords } from "./vizMappings/tree";
+import { canTreeViz, treeFromRecords } from "./visualizations/tree";
 import { BareTerm } from "./replViews";
 
 export type TableCollapseState = {
