+++ uiCommon/visualizations/typeSpec.ts
@@ -1,0 +1,8 @@
+import { Interpreter } from "../../core/interpreter";
+import { Rec } from "../../core/types";
+
+export type VizTypeSpec = {
+  name: string;
+  description: string;
+  component: (props: { interp: Interpreter; spec: Rec }) => React.ReactElement;
+};
