+++ uiCommon/vizRegistry.ts
@@ -1,0 +1,6 @@
+import { VizTypeSpec } from "./visualizations/typeSpec";
+import { sequence } from "./visualizations/sequence";
+
+export const VIZ_REGISTRY: { [id: string]: VizTypeSpec } = {
+  sequence,
+};
